Desktop Client/QR Generator/requirements.txt
customtkinter==5.2.2
pillow==10.3.0
qrcode==7.4.2
